Minus depth from win score
Adding 'minus depth' to the win checker. This way the game will favour moves that win more quickly (a win in 2 moves :8, is worth more than a win in 3 :7). This way if there is a winning move the AI will always choose it, instead of other moves that could eventually win that it happens to find first.